goofy_project/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-guest-user==0.5.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.0
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic-settings==2.3.0
pydantic_core==2.18.4
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
whitenoise==6.6.0
